fix: close six adversarial-audit findings; unify the JSON contract (0.19.0)
A deep adversarial pass surfaced one CI-breaking correctness defect, one
disk-corruption edge, one false positive, one silent runtime skip, three
naming-drift contract inconsistencies, and the last config-over-code
violation. Each is fixed at the root with no compatibility shim.

- numbering check <-> scaffold self-consistency: both derive the sequence
  number through one shared `numbering_sequence` helper -- filter by the
  per-glob `pattern`, parse `u64` -- so a date-prefixed sibling can no
  longer trip a false unique_numbering/sequential_numbering (it broke CI
  under the shipped example pattern). ViolationDetails numbering fields
  widen u32->u64 to match.
- impact dangling-referrer id-match now gates on relation resolution mode
  (is_document_ref_relation), so a path-only `covers` token coinciding
  with a removed id is no longer a false "likely breaking".
- FrontmatterEditor classifies a folded multi-line plain scalar as
  NonScalar (one shared indented-body walk), so retarget/lifecycle refuse
  it instead of writing truncated/invalid YAML to disk.
- cross_field `when` equals/in values are validated at load against the
  field's effective vocabulary/type (dates canonicalized), closing the
  last silent runtime skip: a typo'd predicate is a load error, not a
  rule that never fires.
- naming unification across the JSON contract: SimilarityEntry.similarity
  -> score (all three ranking entries share `score`), IdStability serde
  tag "kind" -> "type" (freeing the `kind` domain term), and the cycle
  built-in id graph_invariants/cycle-detection -> acyclic_relation
  (issue category violation_acyclic_relation).
- scope.prune_dirs lifts the hardcoded NON_CONTENT_DIRS prune list into
  config (default preserves current behavior); scope-coverage warnings
  name a pruning dir as the cause of an unmatched include.

Breaking envelope-schema changes (field/tag/id renames, u64 widening)
-> minor version bump 0.18.0 -> 0.19.0.

+++ nodex-core/src/rules/detail.rs
@@ -142,10 +142,11 @@ pub enum ViolationDetails {
     },
     /// A filename does not match its configured pattern.
     FilenamePattern { filename: String, pattern: String },
-    /// A gap in a directory's sequential file numbering.
-    SequentialNumbering { previous: u32, current: u32 },
+    /// A gap in a directory's sequential file numbering. The number width
+    /// matches `scaffold`'s `u64` sequence so write and check agree.
+    SequentialNumbering { previous: u64, current: u64 },
     /// Two files in a directory share the same number.
-    UniqueNumbering { number: u32, paths: Vec<String> },
+    UniqueNumbering { number: u64, paths: Vec<String> },
     /// A captured body-line token is outside its declared enum.
     BodyLine {
         line: usize,
